build: don't fail on missing SCM directories/files
If a SCM is changed/deleted the offending file or directory is moved
away into the attic. If the old entry was missing this lead to an
internal exception. Now we're checking first if the item is still there
before trying to move it away.

+++ pym/bob/cmds/build.py
@@ -608,12 +608,13 @@ esac
                 for (scmDir, scmDigest) in oldCheckoutState.copy().items():
                     if (scmDir is not None) and (scmDigest != checkoutState.get(scmDir)):
                         scmPath = os.path.normpath(os.path.join(prettySrcPath, scmDir))
-                        atticName = datetime.datetime.now().isoformat()+"_"+os.path.basename(scmPath)
-                        print(colorize("   ATTIC     {} (move to ../attic/{})".format(scmPath, atticName), "33"))
-                        atticPath = os.path.join(prettySrcPath, "..", "attic")
-                        if not os.path.isdir(atticPath):
-                            os.makedirs(atticPath)
-                        os.rename(scmPath, os.path.join(atticPath, atticName))
+                        if os.path.exists(scmPath):
+                            atticName = datetime.datetime.now().isoformat()+"_"+os.path.basename(scmPath)
+                            print(colorize("   ATTIC     {} (move to ../attic/{})".format(scmPath, atticName), "33"))
+                            atticPath = os.path.join(prettySrcPath, "..", "attic")
+                            if not os.path.isdir(atticPath):
+                                os.makedirs(atticPath)
+                            os.rename(scmPath, os.path.join(atticPath, atticName))
                         del oldCheckoutState[scmDir]
                         BobState().setDirectoryState(prettySrcPath, oldCheckoutState)
